% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{methods}
\name{annotation,SNPhood-method}
\alias{annotation}
\alias{annotation,SNPhood-method}
\title{Retrieve the annotation of a \code{SNPhood} object.}
\usage{
\S4method{annotation}{SNPhood}(object, elements = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{SNPhood}}

\item{elements}{Character. The name of the element(s) in the annotation slot to be extracted. 
If set to \code{NULL}, the full annotation slot is returned.}

\item{...}{not supported}
}
\value{
If only a single value for \code{elements} is provided, the element is returned directly. 
If multiple values are provided, a named list with the requested elements is returned.
}
\description{
Specific elements within the annotation slot may also be extracted by using the \code{elements} parameter.
}
\examples{
data(SNPhood.o, package="SNPhood")
annotation(SNPhood.o)
annotation(SNPhood.o, elements = "regions")
annotation(SNPhood.o, elements = c("regions", "bins"))
}

