% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{annotationDatasets}
\alias{annotationDatasets}
\title{Get the annotation(names) of the datasets in a \emph{SNPhood} object.}
\usage{
annotationDatasets(SNPhood.o)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}
}
\value{
Character vector. Names of the datasets/individuals that are defined in the \code{\linkS4class{SNPhood}} object.
}
\description{
Return the names of the datasets/individuals that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
annotationDatasets(SNPhood.o)
}

