% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nBins}
\alias{bins}
\alias{nBins}
\title{Get the number of bins for a \emph{SNPhood} object.}
\usage{
nBins(SNPhood.o, verbose = FALSE)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}

\item{verbose}{Logical(1). Default FALSE. Should the verbose mode (i.e., diagnostic messages during execution of the script) be enabled?}
}
\value{
Integer. Number of bins that are defined in the \code{\linkS4class{SNPhood}} object
}
\description{
Return the number of bins that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
nBins(SNPhood.o)
}

