% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{nDatasets}
\alias{datasets}
\alias{nDatasets}
\title{Get the number of datasets for a \emph{SNPhood} object.}
\usage{
nDatasets(SNPhood.o)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}
}
\value{
Integer. Number of datasets that are defined in the \code{\linkS4class{SNPhood}} object
}
\description{
Return the number of datasets that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
nDatasets(SNPhood.o)
}

