% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,SNPhood-method}
\title{Retrieve the parameters of an object.}
\usage{
parameters(object, ...)

\S4method{parameters}{SNPhood}(object, ...)
}
\arguments{
\item{object}{An object containing parameters with which it was created.}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
A named list with all parameters and its current values of the \code{SNPhood} object.
}
\description{
Retrieve the parameters of an object.

Retrieve the parameters of a \code{SNPhood} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
parameters(SNPhood.o)
}

