% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nRegions}
\alias{nRegions}
\alias{regions}
\title{Get the number of SNP regions for a \emph{SNPhood} object.}
\usage{
nRegions(SNPhood.o, verbose = FALSE)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}

\item{verbose}{Logical(1). Default FALSE. Should the verbose mode (i.e., diagnostic messages during execution of the script) be enabled?}
}
\value{
Integer. Number of SNP regions that are defined in the \code{\linkS4class{SNPhood}} object
}
\description{
\code{nRegions} is a helper function that returns the number of SNP regions that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
nRegions(SNPhood.o)
}

