% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPhood.R
\docType{data}
\name{SNPhood.o}
\alias{SNPhood-data}
\alias{SNPhood.o}
\title{SNPhood example data}
\value{
an example \code{SNPhood} object from the \code{SNPhoodData} package with read counts for 
174 genomic regions across 2 datasets, three read groups and 100 bins
}
\description{
This dataset is an example dataset that can be used for exploring the \code{SNPhood} package.
For more information, see the workflow vignette of the \code{SNPhood} and \code{SNPhoodData} package, respectively.
}
\keyword{datasets}

