% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{annotationBins}
\alias{annotationBins}
\title{Get the annotation(names) of the bins in a \emph{SNPhood} object.}
\usage{
annotationBins(SNPhood.o, verbose = FALSE)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}

\item{verbose}{Logical(1). Default FALSE. Should the verbose mode (i.e., diagnostic messages during execution of the script) be enabled?}
}
\value{
Character vector. Names of the bins that are defined in the \code{\linkS4class{SNPhood}} object.
}
\description{
Return the names of the Bins that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
annotationReadGroups(SNPhood.o)
}

