% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{annotationReadGroups}
\alias{annotationReadGroups}
\title{Get the annotation(names) of the read groups in a \emph{SNPhood} object.}
\usage{
annotationReadGroups(SNPhood.o)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}
}
\value{
Character vector. Names of the read groups that are defined in the \code{\linkS4class{SNPhood}} object.
}
\description{
Return the names of the read groups that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
annotationReadGroups(SNPhood.o)
}

