% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nReadGroups}
\alias{nReadGroups}
\alias{readGroups}
\title{Get the number of read groups for a \emph{SNPhood} object.}
\usage{
nReadGroups(SNPhood.o, verbose = FALSE)
}
\arguments{
\item{SNPhood.o}{Object of class \code{\linkS4class{SNPhood}}}

\item{verbose}{Logical(1). Default FALSE. Should the verbose mode (i.e., diagnostic messages during execution of the script) be enabled?}
}
\value{
Integer. Number of read groups that are defined in the \code{\linkS4class{SNPhood}} object
}
\description{
Return the number of read groups that are defined in the \code{\linkS4class{SNPhood}} object.
}
\examples{
data(SNPhood.o, package="SNPhood")
nReadGroups(SNPhood.o)
}

